requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-filter==25.1
django-redis==5.4.0
django-rest-passwordreset==1.5.0
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Faker==37.0.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
psycopg==3.2.6
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
whitenoise==6.9.0
